# API_SPEC.yaml
openapi: 3.1.0
info:
  title: Galvana Electrochemistry Platform API
  description: |
    Enterprise-grade phygital electrochemistry platform combining real physics simulation 
    with live instrument control, calibration, and data management.
  version: 1.0.0
  contact:
    email: engineering@galvana.com
  license:
    name: Proprietary
servers:
  - url: https://api.galvana.com/v1
    description: Production
  - url: https://api.staging.galvana.com/v1
    description: Staging
  - url: http://localhost:8080/api/v1
    description: Local Development
security:
  - bearerAuth: []
  - apiKey: []
...
paths:
  /runs:
    post:
      tags: [runs]
      summary: Start a new simulation or experiment run
      operationId: createRun
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/SimulationRunRequest'
                - $ref: '#/components/schemas/ExperimentRunRequest'
            examples:
              simulation:
                summary: Start simulation
                value:
                  type: simulation
                  scenario_id: scn_ni_plating_v2
                  engine: fenicsx
                  compute:
                    cpu: 4
                    memory: 16GB
              experiment:
                summary: Start experiment
                value:
                  type: experiment
                  protocol_id: prt_cv_standard
                  instrument_id: ins_gamry_001
                  safety_limits:
                    max_voltage: 5.0
                    max_current: 1.0

      responses:
        '202':
          description: Run accepted and queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunHandle'
          headers:
            Location:
              schema:
                type: string
              description: URL of the created run
            X-Estimated-Duration:
              schema:
                type: integer
              description: Estimated duration in seconds

        '400':
          $ref: '#/components/responses/ValidationError'
        '402':
          $ref: '#/components/responses/QuotaExceeded'
        '503':
          $ref: '#/components/responses/ResourcesUnavailable'

    get:
      tags: [runs]
      summary: List runs with filtering
      operationId: listRuns
      parameters:
        - name: status
          in: query
          schema:
            type: array
            items:
              $ref: '#/components/schemas/RunStatus'
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/RunType'
        - name: created_after
          in: query
          schema:
            type: string
            format: date-time
        - name: created_before
          in: query
          schema:
            type: string
            format: date-time
        - name: tags
          in: query
          schema:
            type: array
            items:
              type: string
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/cursor'

      responses:
        '200':
          description: Paginated list of runs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunList'
  /runs/{runId}:
    get:
      tags: [runs]
      summary: Get run details
      operationId: getRun
      parameters:
        - $ref: '#/components/parameters/runId'
        - name: include
          in: query
          schema:
            type: array
            items:
              type: string
              enum: [scenario, datasets, artifacts, metrics]

      responses:
        '200':
          description: Run details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [runs]
      summary: Update run (pause/resume/abort)
      operationId: updateRun
      parameters:
        - $ref: '#/components/parameters/runId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                action:
                  type: string
                  enum: [pause, resume, abort]
                reason:
                  type: string
                  description: Required for abort action

      responses:
        '200':
          description: Run updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
        '409':
          $ref: '#/components/responses/InvalidStateTransition'
  /runs/{runId}/stream:
    get:
      tags: [runs]
      summary: Stream real-time run data
      operationId: streamRun
      parameters:
        - $ref: '#/components/parameters/runId'
        - name: channels
          in: query
          required: true
          schema:
            type: array
            items:
              type: string
              enum: [status, frames, logs, events, metrics]
        - name: from
          in: query
          schema:
            type: string
            format: date-time
          description: Start streaming from this timestamp

      responses:
        '200':
          description: Server-Sent Events stream
          content:
            text/event-stream:
              schema:
                type: object
                properties:
                  event:
                    type: string
                    enum: [frame, status, log, event, metric]
                  data:
                    type: object
                  id:
                    type: string
                  retry:
                    type: integer
...
  /scenarios:
    post:
      tags: [scenarios]
      summary: Create or update scenario
      operationId: createScenario
      requestBody:
        required: true
        content:
          application/yaml:
            schema:
              $ref: '#/components/schemas/ScenarioYAML'
          application/json:
            schema:
              $ref: '#/components/schemas/Scenario'

      responses:
        '201':
          description: Scenario created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scenario'
          headers:
            Location:
              schema:
                type: string

        '400':
          $ref: '#/components/responses/ValidationError'

    get:
      tags: [scenarios]
      summary: List scenarios
      operationId: listScenarios
      parameters:
        - name: template
          in: query
          schema:
            type: boolean
          description: Filter for template scenarios
        - name: tags
          in: query
          schema:
            type: array
            items:
              type: string
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/cursor'

      responses:
        '200':
          description: List of scenarios
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScenarioList'
  /scenarios/{scenarioId}:
    get:
      tags: [scenarios]
      summary: Get scenario details
      operationId: getScenario
      parameters:
        - $ref: '#/components/parameters/scenarioId'
        - name: format
          in: query
          schema:
            type: string
            enum: [json, yaml]
            default: json

      responses:
        '200':
          description: Scenario details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scenario'
            application/yaml:
              schema:
                $ref: '#/components/schemas/ScenarioYAML'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    runId:
      name: runId
      in: path
      required: true
      schema:
        type: string
        pattern: ^run_[a-zA-Z0-9]+$
    scenarioId:
      name: scenarioId
      in: path
      required: true
      schema:
        type: string
        pattern: ^scn_[a-zA-Z0-9]+$
...
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    cursor:
      name: cursor
      in: query
      schema:
        type: string
  schemas:
    RunType:
      type: string
      enum: [simulation, experiment, hybrid]
    RunStatus:
      type: string
      enum: [queued, starting, running, paused, completed, failed, aborted]
    SimulationRunRequest:
      type: object
      required: [type, scenario]
      properties:
        type:
          type: string
          enum: [simulation]
        scenario:
          oneOf:
            - type: string
              description: Scenario ID
            - $ref: '#/components/schemas/ScenarioYAML'
        engine:
          type: string
          enum: [auto, fenicsx, moose]
          default: auto
        compute:
          $ref: '#/components/schemas/ComputeRequirements'
        tags:
          type: array
          items:
            type: string
        metadata:
          type: object
          additionalProperties: true
    ExperimentRunRequest:
      type: object
      required: [type, protocol, instrument_id]
      properties:
        type:
          type: string
          enum: [experiment]
        protocol:
          oneOf:
            - type: string
              description: Protocol ID
            - $ref: '#/components/schemas/ExperimentProtocol'
        instrument_id:
          type: string
        safety_limits:
          $ref: '#/components/schemas/SafetyLimits'
        tags:
          type: array
          items:
            type: string
    RunHandle:
      type: object
      required: [run_id, status]
      properties:
        run_id:
          type: string
        status:
          $ref: '#/components/schemas/RunStatus'
        queue_position:
          type: integer
        estimated_start:
          type: string
          format: date-time
        estimated_duration:
          type: integer
          description: Seconds
        stream_url:
          type: string
          format: uri
    Run:
      type: object
      required: [id, type, status, created_at, updated_at]
      properties:
        id:
          type: string
        type:
          $ref: '#/components/schemas/RunType'
        status:
          $ref: '#/components/schemas/RunStatus'
        scenario:
          $ref: '#/components/schemas/Scenario'
        progress:
          type: object
          properties:
            percentage:
              type: number
              minimum: 0
              maximum: 100
            current_time:
              type: number
            total_time:
              type: number
            timestep:
              type: integer
        metrics:
          type: object
          properties:
            cpu_usage:
              type: number
            memory_usage:
              type: number
            solver_iterations:
              type: integer
            convergence:
              type: number
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
    RunList:
      type: object
      required: [runs, pagination]
      properties:
        runs:
          type: array
          items:
            $ref: '#/components/schemas/Run'
        pagination:
          $ref: '#/components/schemas/Pagination'
    Scenario:
      type: object
      required: [id, name, version, physics, geometry]
      properties:
        id:
          type: string
        name:
          type: string
        version:
          type: string
        description:
          type: string
        physics:
          $ref: '#/components/schemas/PhysicsConfig'
        geometry:
          $ref: '#/components/schemas/GeometryConfig'
        materials:
          $ref: '#/components/schemas/MaterialsConfig'
        boundaries:
          $ref: '#/components/schemas/BoundariesConfig'
        drive:
          $ref: '#/components/schemas/DriveConfig'
        numerics:
          $ref: '#/components/schemas/NumericsConfig'
        outputs:
          $ref: '#/components/schemas/OutputsConfig'
        tags:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ScenarioYAML:
      type: string
      description: YAML representation of scenario
    PhysicsConfig:
      type: object
      properties:
        transport:
          type: string
          enum: [nernst_planck, stefan_maxwell]
        electroneutral:
          type: boolean
        potential_model:
          type: string
          enum: [poisson, none]
        heat:
          type: boolean
        phase_field:
          type: boolean
    GeometryConfig:
      type: object
      properties:
        type:
          type: string
          enum: [1D, 2D, 3D]
        dimensions:
          type: object
        mesh:
          type: object
          properties:
            elements:
              type: integer
            refinement:
              type: object
    MaterialsConfig:
      type: object
      properties:
        electrolyte:
          type: object
          properties:
            species:
              type: array
              items:
                type: object
                properties:
                  name:
                    type: string
                  diffusivity:
                    type: number
                  charge:
                    type: integer
            conductivity:
              type: number
            activity_model:
              type: string
        electrode:
          type: object
          properties:
            porosity:
              type: number
            tortuosity:
              type: number
    BoundariesConfig:
      type: object
      additionalProperties:
        type: object
        properties:
          type:
            type: string
            enum: [electrode, counter, insulation, periodic]
          reaction:
            type: string
          potential:
            type: number
    DriveConfig:
      type: object
      properties:
        mode:
          type: string
          enum: [potentiostatic, galvanostatic, potentiodynamic]
        waveform:
          $ref: '#/components/schemas/Waveform'
    Waveform:
      type: object
      properties:
        type:
          type: string
          enum: [step, ramp, sine, triangle, cv, pulse]
        parameters:
          type: object
          additionalProperties: true
    NumericsConfig:
      type: object
      properties:
        time_integrator:
          type: string
          enum: [BDF, SDIRK, explicit_euler]
        dt_initial:
          type: number
        dt_max:
          type: number
        newton_tol:
          type: number
        linear_solver:
          type: string
          enum: [gmres, bicgstab, direct]
    OutputsConfig:
      type: object
      properties:
        save:
          type: array
          items:
            type: string
        cadence:
          type: number
        format:
          type: string
          enum: [hdf5, netcdf, csv, zarr]
    ComputeRequirements:
      type: object
      properties:
        cpu:
          type: integer
          minimum: 1
        memory:
          type: string
          pattern: ^\d+[MG]B$
        gpu:
          type: boolean
        priority:
          type: string
          enum: [low, normal, high]
    SafetyLimits:
      type: object
      properties:
        max_voltage:
          type: number
        max_current:
          type: number
        max_power:
          type: number
        max_temperature:
          type: number
        emergency_contacts:
          type: array
          items:
            type: string
...
    ScenarioList:
      type: object
      properties:
        scenarios:
          type: array
          items:
            $ref: '#/components/schemas/Scenario'
        pagination:
          $ref: '#/components/schemas/Pagination'
    ExperimentProtocol:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [cv, ca, cp, eis, gitt, ocv]
        parameters:
          type: object
          additionalProperties: true
...
    Pagination:
      type: object
      properties:
        limit:
          type: integer
        cursor:
          type: string
        has_more:
          type: boolean
        total_count:
          type: integer
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
            request_id:
              type: string
            documentation_url:
              type: string
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: VALIDATION_ERROR
              message: Invalid scenario configuration
              details:
                field: physics.transport
                reason: Invalid transport model
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    QuotaExceeded:
      description: Quota or rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      headers:
        Retry-After:
          schema:
            type: integer
    ResourcesUnavailable:
      description: Compute resources unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InvalidStateTransition:
      description: Invalid state transition
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...